FE reservation logic and auth improvement

+++ packages/server/src/features/parking-reservations/services/parking-reservations.service.ts
@@ -6,6 +6,8 @@ import { PARKING_RESERVATION_SPACES } from '../../../shared/constants/constant';
 import { NOT_ALLOWED_MESSAGES, NOT_FOUND_MESSAGES } from '../../../shared/enums/texts.enum';
 import { CreateParkingReservationDto } from '../dto/create-parking-reservation.dto';
 import { UpdateParkingReservationDto } from '../dto/update-parking-reservation.dto';
+import { User } from '../../../schemas/user.schema';
+import { Role } from '../../../shared/enums/auth.enum';
 
 @Injectable()
 export class ParkingReservationsService {
@@ -70,11 +72,23 @@ export class ParkingReservationsService {
     return updatedReservation;
   }
 
-  async remove(_id: string): Promise<ParkingReservation> {
+  async remove(_id: string, user: User): Promise<ParkingReservation> {
+    const isCreator = this.isCreator(_id, user);
+
+    if (!isCreator) throw new MethodNotAllowedException(NOT_ALLOWED_MESSAGES.USER_IS_NOT_CREATOR);
+
     const deletedReservation = await this._parkingReservationModel.findByIdAndDelete({ _id }).exec();
 
     if (!deletedReservation) throw new NotFoundException(NOT_FOUND_MESSAGES.PARKING_RESERVATION_NOT_FOUND);
 
     return deletedReservation;
+  }
+
+  async isCreator(_id: string, user: User): Promise<boolean> {
+    if (!(user.role === Role.ADMIN)) {
+      const reservation = await this.findOne(_id);
+      if (reservation.userId.toString() !== user._id.toString()) return false;
+    }
+    return true;
   }
 }
